fix(minor): minor fix for tempat tinggal and seeder

+++ database/migrations/2024_11_30_143556_create_tempat_tinggal_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('kodePos', 100);
             $table->string('provinsi', 100);
             $table->string('tinggalBersama', 100);
-            $table->integer('jarakKeSekolah');
+            $table->float('jarakKeSekolah');
             $table->string('kendaraan', 100);
             $table->timestamps();
